backend/requirements.txt
# 中醫螺旋案例推理輔助診斷系統 - 依賴套件

# 核心依賴
pyyaml>=6.0
python-dotenv>=1.0.0

# Web API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0 # Note: pydantic-settings is used for BaseSettings
python-dotenv>=1.0.1
# Weaviate Client (v4 compatible)
weaviate-client>=4.4.0,<5.0.0
openai>=1.12.0
numpy>=1.26.0
scikit-learn>=1.4.0

# HTTP Client
httpx>=0.26.0 # For Nvidia Client

# Templating Engine
Jinja2>=3.1.3 # For Report Generation